fim do dia 3

+++ app/Http/Controllers/ExportController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Export;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Inertia\Inertia;
 
 class ExportController extends Controller
 {
@@ -15,20 +16,28 @@ class ExportController extends Controller
      */
     public function index()
     {
-        $exports = Export::paginate(15);
+        $exports = Export::paginate(5);
+
+        return Inertia::render('Reports', [
+            'exports' => $exports
+        ]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Export  $export
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Export $export)
-    {
-        Storage::delete($export->file_name);
-        $export->delete();
+    public function show($export) {
+        $export = Export::find($export);
+        
+        return Storage::download($export->file_name);
+    }
 
-        return "deletando";
+    public function destroy($export)
+    {
+        $export = Export::find($export);
+
+        if ($export) {
+            Storage::delete($export->file_name);
+            $export->delete();
+        }
+
+        return redirect()->back()->with('success', 'Seu arquivo foi removido com sucesso!');
     }
 }
